Revert "Move TensorOptions to ATen/core (#22020)"
This reverts commit bde73860c6c88cb21dda0ba2cbbc3ce7c026b6fd.
