+++ aten/src/ATen/InitialTensorOptions.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 
 namespace at {
