+++ aten/src/ATen/native/quantized/TensorFactories.cpp
@@ -1,8 +1,7 @@
 #include <ATen/ATen.h>
-
-#include <ATen/core/TensorOptions.h>
 #include <ATen/quantized/Quantizer.h>
 #include <c10/core/QScheme.h>
+#include <c10/core/TensorOptions.h>
 
 namespace at {
 namespace native {
