+++ aten/src/ATen/quantized/Quantizer.h
@@ -6,7 +6,7 @@
 #include <c10/util/Exception.h>
 #include <c10/util/intrusive_ptr.h>
 #include <c10/core/ScalarType.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 
 #include <cmath>
 #include <memory>
