+++ aten/src/ATen/templates/Functions.h
@@ -9,7 +9,7 @@
 #include <c10/util/Deprecated.h>
 #include <ATen/NativeFunctions.h>
 #include <ATen/DeviceGuard.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 #include <ATen/core/Reduction.h>
 #include <c10/util/Optional.h>
 #include <ATen/TensorUtils.h>
