+++ aten/src/ATen/templates/TypeDefault.cpp
@@ -12,7 +12,7 @@
 #include <c10/core/Scalar.h>
 #include <c10/core/Storage.h>
 #include <ATen/Tensor.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 #include <ATen/DeviceGuard.h>
 #include <ATen/SparseTensorUtils.h>
 #include <ATen/core/ATenDispatch.h>
