+++ test/cpp/api/tensor_options.cpp
@@ -4,7 +4,7 @@
 
 #include <ATen/Context.h>
 #include <ATen/Functions.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 
 #include <string>
 #include <vector>
