+++ torch/csrc/api/include/torch/nn/cloneable.h
@@ -4,7 +4,7 @@
 #include <torch/types.h>
 #include <torch/utils.h>
 
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 #include <c10/util/Exception.h>
 
 #include <memory>
