+++ torch/csrc/api/include/torch/nn/parallel/data_parallel.h
@@ -14,7 +14,7 @@
 
 #include <ATen/Device.h>
 #include <ATen/Parallel.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 #include <c10/util/Exception.h>
 
 #include <cstddef>
