+++ torch/csrc/autograd/python_variable_indexing.cpp
@@ -15,7 +15,7 @@
 
 #include <ATen/DeviceGuard.h>
 #include <ATen/ExpandUtils.h>
-#include <ATen/core/TensorOptions.h>
+#include <c10/core/TensorOptions.h>
 #include <ATen/core/LegacyTypeDispatch.h>
 
 #include <vector>
